Moving inventory_provider to controller
