--- suzieq/inventory_provider/examples/example-config.yml
@@ -1,16 +1,0 @@
-inventory_provider:
-  plugin_type:
-    inventory_source:
-      - source_path: "suzieq/inventory_provider/examples/inventory.yml"
-    poller_manager:
-      - type: static_poller_manager
-        pollers_count: 1
-        inventory_path: /home/luca/stardust/suzieq
-        inventory_file_name: static_inv
-        start_pollers: false
-    chunker:
-      - type: static_chunker
-        split_pol: namespace
-  provider_config:
-    period: 10
-    inventory_get_timeout: 20
