--- suzieq/inventory_provider/plugins/inventory_source/examples/local-netbox-devices-credentials.yaml
@@ -1,21 +1,0 @@
-namespace:
-  - name: "local"
-    devices:
-      - name: TDevice0
-        credentials:
-          username: username
-          password: password
-        options:
-          ask_pass: True
-          envpass: TDevice0_PASSWORD
-          # ...
-      - name: TDevice1
-        credentials:
-          key_path: "path/to/key/file"
-          username: username
-        # password: password
-        options:
-          ask_pass: True
-          envpass: TDevice1_PASSWORD
-          passphrase: suzieqSuzieq
-          # ...
