--- suzieq/inventory_provider/tests/config/netbox1.yaml
@@ -1,6 +1,0 @@
-name: netbox1
-server_info:
-  port: 9000
-  token: "12345678901234567890"
-  username: admin
-  password: admin
