--- suzieq/inventory_provider/tests/config/netbox2.yaml
@@ -1,6 +1,0 @@
-name: netbox2
-server_info:
-  port: 9001
-  token: "12345678901234567890"
-  username: admin
-  password: admin
