--- suzieq/inventory_provider/tests/generators/inventory_source/netbox_generator/api/users/tokens/provision/netbox2.json
@@ -1,1 +1,0 @@
-[{"id": 1, "url": "https://demo.netbox.dev/api/users/tokens/1/", "display": "002963 (admin)", "user": {"id": 1, "url": "https://demo.netbox.dev/api/users/users/1/", "display": "admin", "username": "admin"}, "created": "2021-11-25T10:36:46.601998Z", "expires": null, "key": "0123456789123456789", "write_enabled": true, "description": ""}, {"id": 1, "url": "https://demo.netbox.dev/api/users/tokens/1/", "display": "002963 (admin)", "user": {"id": 1, "url": "https://demo.netbox.dev/api/users/users/1/", "display": "admin", "username": "admin"}, "created": "2021-11-25T10:36:46.601998Z", "expires": null, "key": "akjasdpksdjcnaspkdjnfa", "write_enabled": true, "description": ""}]
