--- suzieq/inventory_provider/tests/generators/inventory_source/netbox_generator/common/base_device.json
@@ -1,88 +1,0 @@
-{
-    "id": 1,
-    "url": "http://127.0.0.1:8000/api/dcim/devices/1/",
-    "display": "TDevice0",
-    "name": "TDevice0",
-    "device_type": {
-        "id": 1,
-        "url": "http://127.0.0.1:8000/api/dcim/device-types/1/",
-        "display": "IT-19129381",
-        "manufacturer": {
-            "id": 1,
-            "url": "http://127.0.0.1:8000/api/dcim/manufacturers/1/",
-            "display": "MSI",
-            "name": "MSI",
-            "slug": "msi"
-        },
-        "model": "IT-19129381",
-        "slug": "it-19129381"
-    },
-    "device_role": {
-        "id": 1,
-        "url": "http://127.0.0.1:8000/api/dcim/device-roles/1/",
-        "display": "TestingDevice",
-        "name": "TestingDevice",
-        "slug": "testingdevice"
-    },
-    "tenant": null,
-    "platform": null,
-    "serial": "000000000000000",
-    "asset_tag": null,
-    "site": {
-        "id": 1,
-        "url": "http://127.0.0.1:8000/api/dcim/sites/1/",
-        "display": "TestingSite",
-        "name": "TestingSite",
-        "slug": "testingsite"
-    },
-    "location": null,
-    "rack": null,
-    "position": null,
-    "face": null,
-    "parent_device": null,
-    "status": {
-        "value": "active",
-        "label": "Active"
-    },
-    "primary_ip": {
-        "id": 2,
-        "url": "http://127.0.0.1:8000/api/ipam/ip-addresses/2/",
-        "display": "10.0.0.2/24",
-        "family": 4,
-        "address": "10.0.0.2/24"
-    },
-    "primary_ip4": {
-        "id": 2,
-        "url": "http://127.0.0.1:8000/api/ipam/ip-addresses/2/",
-        "display": "10.0.0.2/24",
-        "family": 4,
-        "address": "10.0.0.2/24"
-    },
-    "primary_ip6": {
-        "id": 3,
-        "url": "http://127.0.0.1:8000/api/ipam/ip-addresses/3/",
-        "display": "6bf0:ed49:b2aa:626a:3190:912f:a136:14ae/128",
-        "family": 6,
-        "address": "6bf0:ed49:b2aa:626a:3190:912f:a136:14ae/128"
-    },
-    "cluster": null,
-    "virtual_chassis": null,
-    "vc_position": null,
-    "vc_priority": null,
-    "comments": "",
-    "local_context_data": null,
-    "tags": [
-        {
-            "id": 1,
-            "url": "http://127.0.0.1:8000/api/extras/tags/1/",
-            "display": "suzieq",
-            "name": "suzieq",
-            "slug": "suzieq",
-            "color": "673ab7"
-        }
-    ],
-    "custom_fields": {},
-    "config_context": {},
-    "created": "2021-11-11",
-    "last_updated": "2021-11-17T14:54:43.639995Z"
-}
