--- suzieq/inventory_provider/tests/generators/inventory_source/netbox_generator/common/base_token.json
@@ -1,16 +1,0 @@
-{
-    "id": 1,
-    "url": "https://demo.netbox.dev/api/users/tokens/1/",
-    "display": "002963 (admin)",
-    "user": {
-        "id": 1,
-        "url": "https://demo.netbox.dev/api/users/users/1/",
-        "display": "admin",
-        "username": "admin"
-    },
-    "created": "2021-11-25T10:36:46.601998Z",
-    "expires": null,
-    "key": "99499204e7a484d0825d962ae5225fb27f002963",
-    "write_enabled": true,
-    "description": ""
-}
