--- suzieq/inventory_provider/tests/generators/inventory_source/netbox_generator/errors/403.invalid_credentials
@@ -1,1 +1,0 @@
-Invalid username/password
