--- suzieq/inventory_provider/tests/generators/inventory_source/netbox_generator/errors/403.invalid_token
@@ -1,1 +1,0 @@
-Invalid token
