--- suzieq/inventory_provider/tests/generators/inventory_source/netbox_generator/errors/403.no_token
@@ -1,1 +1,0 @@
-Authentication credentials were not provided.
