--- suzieq/inventory_provider/tests/generators/inventory_source/netbox_generator/errors/404.page_not_found
@@ -1,1 +1,0 @@
-Page not found
