--- suzieq/inventory_provider/tests/generators/inventory_source/netbox_generator/example/REST_config.yaml
@@ -1,3 +1,0 @@
-netbox_name: netbox1
-ip: 127.0.0.1
-port: 9000
