+++ suzieq/poller/controller/credential_loader/credential_loader.py
@@ -1,0 +1,82 @@
+"""This module contains the base class for plugins which loads
+devices credentials
+"""
+from abc import abstractmethod
+from typing import Dict, List, Type
+from suzieq.shared.sq_plugin import SqPlugin
+
+
+class CredentialLoader(SqPlugin):
+    """Base class used to import device credentials from different
+    sources
+    """
+
+    def __init__(self, init_data) -> None:
+        super().__init__()
+
+        self._cred_format = [
+            "username",
+            "password",
+            "ssh_keyfile",
+            "options"
+        ]
+
+        self.init(init_data)
+
+    def init(self, init_data: Type):
+        """Initialize the object
+
+        Args:
+            init_data (Type): data used to initialize the object
+        """
+
+    @abstractmethod
+    def load(self, inventory: Dict[str, Dict]):
+        """Loads the credentials inside the inventory
+
+        Args:
+            inventory (Dict[str, Dict]): inventory to update
+        """
+
+    def write_credentials(self, device: Dict, credentials: Dict[str, Dict]):
+        """write and validate input credentials for a device
+
+        Args:
+            device (Dict): device to add credentials
+            credentials (Dict[str, Dict]): device's credentials
+
+        Raises:
+            ValueError: Invalid credentials
+        """
+        device.update(credentials)
+        # device["credentials"] = credentials
+        missing_keys = self._validate_credentials(device)
+        if missing_keys:
+            raise ValueError(
+                f"Invalid credentials: missing keys {missing_keys}")
+
+    def _validate_credentials(self, device: Dict) -> List[str]:
+        # credentials = device.get("credentials", {})
+        # if not credentials:
+        #     return ["credentials"]
+
+        cred_keys = set(self._cred_format)
+        for key in device.keys():
+            if key in cred_keys:
+                cred_keys.remove(key)
+
+        # One between password or ssh_keyfile must be defines
+        if "password" in cred_keys and "ssh_keyfile" in cred_keys:
+            cred_keys.remove("password")
+            cred_keys.remove("ssh_keyfile")
+            ret = list(cred_keys)
+            ret.append("password or ssh_keyfile")
+            return ret
+
+        if "password" in cred_keys:
+            cred_keys.remove("password")
+
+        if "ssh_keyfile" in cred_keys:
+            cred_keys.remove("ssh_keyfile")
+
+        return list(cred_keys)
