+++ suzieq/poller/controller/examples/example-config.yml
@@ -1,0 +1,16 @@
+controller:
+  plugin_type:
+    source:
+      - source_path: "suzieq/sq-inventory.yml"
+    manager:
+      - type: static_manager
+        pollers_count: 1
+        inventory_path: suzieq
+        inventory_file_name: static_inv
+        start_pollers: true
+    chunker:
+      - type: static_chunker
+        split_pol: namespace
+  controller_config:
+    period: 10
+    timeout: 20
