+++ suzieq/poller/controller/examples/inventory.yml
@@ -1,0 +1,10 @@
+- namespace: "testing"
+  type: netbox
+  token: "af8717c89ec0ff420c19d89e6c20646ad05dd54e"
+  url: "http://127.0.0.1:8000"
+  transport: ssh
+  tag: "suzieq-demo"
+  period: 60
+  device_credentials:
+    type: cred_file
+    file_path: "/home/luca/stardust/random/devices_credentials.yaml"
